Communicate changes made during methods back to the caller
Right now, we have places where we'll make some update to, say
an instance, and each time we do that, we get back a fresh copy
of the instance in case other properties have been changed since
we fetched our object. This change makes the object_action()
return information about fields that were changed during a method,
if the method implementation left them in the what_changed()
list. Only these changes are communicated back to the caller
over RPC and applied to the caller's object.

In local mode, changes made to the object by the method
implementation will, obviously, be visible to the caller, giving
the same semantics as remote.

Change-Id: I7bee4e1b985c583248e1a496379aa17c88f33b39