More instance usage fixes
This implements Instance.refresh() as a lightweight (RPC-wise)
way to refresh your object from the database. This uses the
method-initiated updates function of a couple patches ago, and its
use is demonstrated by removing a bit of redundant instance-fetching
in test_compute.

This is messy because it fixes other issues with instance, but
the meat is in instance_obj.Instance::refresh() and the
corresponding test_object function for it.

Change-Id: Id608c5dd4139950fd25b346bcc53b68b6e698105