Change RPC services to pass context to the serializer functions
This would be a change to Oslo, but I think that code is going to look
differently than it does now anyway. So, do this for the moment. This
will stab _context into any instance on the way into a method via args
or on the way out of the rpc proxy on result.

Change-Id: I1a68e8172aa40594110c601f259c4e294f94435c